+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Changelog
+## v2.0.19.3
+- Make sure the received ad key is an object before get object's first element
+
 ## v2.0.19.2
 - Avoid use enabled property without checking the existence
 - Monitor ad errors in honeybadger tool
 - Fix broken webkitCancelFullScreen call occured on Firefox browser
-
 
 ## v2.0.19.1
 - Fix importing share function depending on presence of any cuepoints